Remove Logging And Config From The System
Does logging really add anything to this project? Not really at the
moment. So lets remove it untill we find its needed latter on.

The config was overly complex for no reason, poorly implemented and
added nothing to the project. So lets just get rid of it, for now at
least.